requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
pytest>=7.2
pytest-django>=4.5
requests>=2.31
python-dotenv>=1.0
django-debug-toolbar>=4.1
pytest-cov>=4.1

